Updated tests for curly brace recursion
