+++ tests/testData/tokenizerErrorData.js
